requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.6
urllib3==1.26.12
